backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
openai>=1.40.0
pillow>=10.3.0
cairosvg>=2.7.1
openpyxl>=3.1.2